+++ Source/Scene/ImageBasedLighting.js
@@ -191,7 +191,7 @@ Object.defineProperties(ImageBasedLighting.prototype, {
    * computed from the atmosphere color is used.
    * <p>
    * There are nine <code>Cartesian3</code> coefficients.
-   * The order of the coefficients is: L<sub>00</sub>, L<sub>1-1</sub>, L<sub>10</sub>, L<sub>11</sub>, L<sub>2-2</sub>, L<sub>2-1</sub>, L<sub>20</sub>, L<sub>21</sub>, L<sub>22</sub>
+   * The order of the coefficients is: L<sub>0,0</sub>, L<sub>1,-1</sub>, L<sub>1,0</sub>, L<sub>1,1</sub>, L<sub>2,-2</sub>, L<sub>2,-1</sub>, L<sub>2,0</sub>, L<sub>2,1</sub>, L<sub>2,2</sub>
    * </p>
    *
    * These values can be obtained by preprocessing the environment map using the <code>cmgen</code> tool of